requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Environment & Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Document Processing
python-docx==1.1.0
pdf2docx==0.5.8
openpyxl==3.1.2

# Image & PDF Processing
Pillow==10.2.0
img2pdf==0.5.1
pdf2image==1.17.0

# Celery & Redis (Distributed Task Queue)
celery[redis]==5.3.6
redis==5.0.3
flower==2.0.1

# Utilities
aiofiles==23.2.1
